add CALPADS schools CSV parsing
