+++ package-processor/src/main/java/org/opentestsystem/rdw/ingest/processor/service/impl/OrganizationParser.java
@@ -1,0 +1,27 @@
+package org.opentestsystem.rdw.ingest.processor.service.impl;
+
+import org.opentestsystem.rdw.ingest.common.model.School;
+import org.opentestsystem.rdw.ingest.common.util.DataElementErrorCollector;
+import org.opentestsystem.rdw.ingest.common.util.ParserHelper;
+
+import java.util.Map;
+
+/**
+ * Parses organization payloads; helper for DefaultOrganizationProcessor
+ */
+public interface OrganizationParser {
+
+    /**
+     * Parse payload and extract schools and districts.<br/>
+     * This method should throw if the payload is unparseable. If the payload is parseable but has errors,
+     * let the parser helper populate the error collector and return without throwing. If the payload is
+     * parseable and good, return with throwing or populating the error collector.
+     *
+     * @param payload payload to parse
+     * @param parserHelper parser helper
+     * @throws IllegalArgumentException if payload is not parseable
+     */
+    void parse(byte[] payload, ParserHelper parserHelper);
+
+    Map<String, School> getSchools();
+}
