+++ package-processor/src/test/resources/CA_schools.csv
@@ -1,0 +1,24 @@
+County-District Code^School Code^Auth CDS Code^County Name^District Name^School Name^Charter School^Charter Status^NPS School
+0100016^0100016^20652430100016^Madera^Sherman Thomas Charter^Sherman Thomas Charter^Y^DF^N
+0100065^0100065^01612590100065^Alameda^Oakland Unity High^Oakland Unity High^Y^DF^N
+0100099^0100099^55723630100099^Tuolumne^California Virtual Academy @ Jamestown^California Virtual Academy @ Jamestown^Y^DF^N
+0100123^0100123^01612590100123^Alameda^East Oakland Leadership Academy^East Oakland Leadership Academy^Y^DF^N
+0561580^6111884^05615806111884^Calaveras^Vallecito Union^Avery Middle^N^^N
+0610066^0610066^06100660610066^Colusa^Colusa County Office of Education School District^Colusa County Office of Education District Level Program^N^^N
+1075275^6112544^10752756112544^Fresno^Sierra Unified^Lodge Pole Elementary (Alternative)^N^^N
+1075408^1030253^10754081030253^Fresno^Riverdale Joint Unified^Horizon High^N^^N
+1262679^6007678^12626796007678^Humboldt^Arcata Elementary^Arcata Elementary^N^^N
+1262687^0107110^12626870107110^Humboldt^Northern Humboldt Union High^Six Rivers Charter High^Y^LF^N
+1964733^6061568^19647336061568^Los Angeles^Los Angeles Unified^Woodland Hills Academy^N^^N
+1964733^6061576^19647336061576^Los Angeles^Los Angeles Unified^Robert E. Peary Middle^N^^N
+1964733^6061584^19647336061584^Los Angeles^Los Angeles Unified^Gaspar De Portola Middle^N^^N
+3768056^6120596^37680566120596^San Diego^Del Mar Union Elementary^Torrey Hills^N^^N
+3768056^6987960^37680566987960^San Diego^Del Mar Union Elementary^T.I.E.E. Site 1 - Children's Workshop (The Institute for Effective Education)^N^^Y
+3768080^3768080^37680803768080^San Diego^Encinitas Union Elementary School District^Encinitas Union Elementary District Level Program^N^^N
+3768080^6038145^37680806038145^San Diego^Encinitas Union Elementary^Paul Ecke-Central Elementary^N^^N
+3768080^6038152^37680806038152^San Diego^Encinitas Union Elementary^Ocean Knoll Elementary^N^^N
+6120893^6120893^37684036120893^San Diego^California Virtual Academy @ San Diego^California Virtual Academy @ San Diego^Y^DF^N
+6120901^6120901^37681896120901^San Diego^Barona Indian Charter^Barona Indian Charter^Y^DF^N
+6120935^6120935^37683386120935^San Diego^Albert Einstein Academy Charter Elementary^Albert Einstein Academy Charter Elementary^Y^DF^N
+6121081^6121081^19647336121081^Los Angeles^View Park Preparatory Accelerated Charter Middle^View Park Preparatory Accelerated Charter Middle^Y^DF^N
+7776422^7776422^77764227776422^^Out-of-State, Non-Public, Non-Sectarian School District^Out-of-State, Non-Public, Non-Sectarian District Level Program^N^^N
