+++ package-processor/src/test/resources/organization.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<root>
+  <!-- this is a test file that is in an unsupported format -->
+</root>
